fix(lint): move hasNonEmptyRows to utils to satisfy import restrictions
