+++ src/components/composite/agGridTable/agGridTable-utils.ts
@@ -6,7 +6,6 @@
  */
 
 import type { ReactNode } from 'react';
-import { FieldConstants } from '../../../utils';
 
 export interface CsvProps {
     fileName: string;
@@ -16,17 +15,3 @@ export interface CsvProps {
     extraButtons?: ReactNode;
     hasTableData?: boolean;
 }
-
-/**
- * Returns true when the given form rows contain at least one cell with a
- * non-empty value (excluding the AG_GRID_ROW_UUID identifier column).
- */
-export const hasNonEmptyRows = (rows: unknown): boolean =>
-    Array.isArray(rows) &&
-    rows.some(
-        (row: any) =>
-            row &&
-            Object.keys(row)
-                .filter((key) => key !== FieldConstants.AG_GRID_ROW_UUID)
-                .some((key) => row[key] !== undefined && row[key] !== null && String(row[key]).trim().length > 0)
-    );
