+++ src/utils/functions.ts
@@ -5,6 +5,22 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
+import { FieldConstants } from './constants/fieldConstants';
+
+/**
+ * Returns true when the given form rows contain at least one cell with a
+ * non-empty value (excluding the AG_GRID_ROW_UUID identifier column).
+ */
+export const hasNonEmptyRows = (rows: unknown): boolean =>
+    Array.isArray(rows) &&
+    rows.some(
+        (row: any) =>
+            row &&
+            Object.keys(row)
+                .filter((key) => key !== FieldConstants.AG_GRID_ROW_UUID)
+                .some((key) => row[key] !== undefined && row[key] !== null && String(row[key]).trim().length > 0)
+    );
+
 /**
  * function to generate a key
  * @returns {number} key
